`@remotion/maptiler`: Add package with MapTiler components
