+++ packages/example/package.json
@@ -31,7 +31,6 @@
 		"render-transparent": "remotion render --image-format=png --pixel-format=yuva420p --codec=vp8 react-svg out/video.webm"
 	},
 	"dependencies": {
-		"@maptiler/sdk": "4.0.2",
 		"@remotion/studio-protocol": "workspace:*",
 		"@mdx-js/loader": "2.3.0",
 		"@mdx-js/react": "2.3.0",
@@ -57,6 +56,7 @@
 		"@remotion/lambda-client": "workspace:*",
 		"@remotion/layout-utils": "workspace:*",
 		"@remotion/light-leaks": "workspace:*",
+		"@remotion/maptiler": "workspace:*",
 		"@remotion/mac-cursors": "workspace:*",
 		"@remotion/rough-notation": "workspace:*",
 		"@remotion/starburst": "workspace:*",
@@ -86,7 +86,6 @@
 		"@remotion/zod-types": "workspace:*",
 		"@remotion/zod-types-v3": "workspace:*",
 		"@shopify/react-native-skia": "2.0.0",
-		"@turf/turf": "7.3.2",
 		"@types/express": "4.17.23",
 		"@types/hls.js": "0.13.3",
 		"@types/opentype.js": "1.3.4",
