+++ packages/maptiler/README.md
@@ -1,0 +1,18 @@
+# @remotion/maptiler
+
+MapTiler components for Remotion
+
+[![NPM Downloads](https://img.shields.io/npm/dm/@remotion/maptiler.svg?style=flat&color=black&label=Downloads)](https://npmcharts.com/compare/@remotion/maptiler?minimal=true)
+
+## Installation
+
+```bash
+npm install @remotion/maptiler --save-exact
+```
+
+When installing a Remotion package, make sure to align the version of all `remotion` and `@remotion/*` packages to the same version.
+Remove the `^` character from the version number to use the exact version.
+
+## Usage
+
+This package is currently internal and has no public documentation yet.
