+++ packages/maptiler/eslint.config.mjs
@@ -1,0 +1,7 @@
+import {remotionFlatConfig} from '@remotion/eslint-config-internal';
+
+const config = remotionFlatConfig({react: true});
+
+export default {
+	...config,
+};
