+++ packages/maptiler/src/index.ts
@@ -1,0 +1,4 @@
+export {MapOverlay} from './MapOverlay';
+export {MapRegion, type MapRegionFeature} from './MapRegion';
+export {MapRoute, type MapRouteFeature} from './MapRoute';
+export {MapViewport} from './MapViewport';
