+++ packages/studio-shared/src/package-info.ts
@@ -99,6 +99,7 @@ const allPackages = [
 	'studio-protocol',
 	'light-leaks',
 	'mac-cursors',
+	'maptiler',
 	'rough-notation',
 	'starburst',
 	'vercel',
@@ -258,6 +259,7 @@ export const descriptions: {[key in Pkgs]: string | null} = {
 		'Create Element payloads and request installation into Remotion Studio',
 	'light-leaks': 'Light leak effects for Remotion',
 	'mac-cursors': 'macOS cursor components for Remotion',
+	maptiler: 'MapTiler components for Remotion',
 	'rough-notation': 'Rough annotation primitives for Remotion',
 	'player-a11y': 'Internal accessibility wrapper around @remotion/player',
 	starburst: 'Starburst ray effect for Remotion',
@@ -375,6 +377,7 @@ export const installableMap: {[key in Pkgs]: boolean} = {
 		packageName: '@remotion/mac-cursors',
 		releaseVersion: VERSION,
 	}),
+	maptiler: false,
 	'rough-notation': true,
 	starburst: shouldReleasePackage({
 		packageName: '@remotion/starburst',
@@ -486,6 +489,7 @@ export const apiDocs: {[key in Pkgs]: string | null} = {
 	'studio-protocol': 'https://www.remotion.dev/docs/studio-protocol',
 	'light-leaks': 'https://www.remotion.dev/docs/light-leaks',
 	'mac-cursors': 'https://www.remotion.dev/docs/mac-cursors',
+	maptiler: null,
 	'rough-notation': 'https://www.remotion.dev/docs/rough-notation',
 	starburst: 'https://www.remotion.dev/docs/starburst',
 	vercel: 'https://www.remotion.dev/docs/vercel/api',
